requirements.txt
asgiref==3.8.1
Django==5.0.7
django-compress-field==0.10.2
django-cors-headers==4.4.0
django-jsonfield==1.4.1
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
phonenumbers==8.13.40
psycopg2-binary==2.9.9
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
